+++ src/components/EtherspotTransactionKit.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { WalletProviderLike, Factory } from '@etherspot/prime-sdk';
+import { WalletProviderLike } from '@etherspot/prime-sdk';
+import { WalletProviderLike as WalletProviderLikeModular} from '@etherspot/modular-sdk';
 
 // types
 import { AccountTemplate } from '../types/EtherspotTransactionKit';
@@ -8,9 +9,8 @@ import { AccountTemplate } from '../types/EtherspotTransactionKit';
 import EtherspotTransactionKitContextProvider from '../providers/EtherspotTransactionKitContextProvider';
 import ProviderWalletContextProvider from '../providers/ProviderWalletContextProvider';
 import EtherspotContextProvider from '../providers/EtherspotContextProvider';
-
 interface EtherspotTransactionKitProps extends React.PropsWithChildren {
-  provider: WalletProviderLike;
+  provider: WalletProviderLike | WalletProviderLikeModular;
   chainId?: number;
   accountTemplate?: AccountTemplate;
   dataApiKey?: string;
@@ -21,16 +21,38 @@ const EtherspotTransactionKit = ({
   children,
   provider,
   chainId = 1,
-  accountTemplate = Factory.ETHERSPOT,
+  accountTemplate,
   dataApiKey,
   bundlerApiKey,
-}: EtherspotTransactionKitProps) => (
+}: EtherspotTransactionKitProps) => {
+  let accountTemp: AccountTemplate;
+
+  switch (accountTemplate) {
+    case 'etherspotModular':
+      accountTemp = 'etherspotModular';
+      break;
+    case 'etherspot':
+      accountTemp = 'etherspot';
+      break;
+    case 'simpleAccount':
+      accountTemp = 'simpleAccount';
+      break;
+    case 'zeroDev':
+      accountTemp = 'zeroDev';
+      break;
+    default:
+      accountTemp = 'etherspotModular';
+      break;
+  }
+
+  return (
   <EtherspotContextProvider
     provider={provider}
     chainId={+chainId} // cast to make it less failproof when passed as string, i.e. from env file
-    accountTemplate={accountTemplate}
+    accountTemplate={accountTemp}
     dataApiKey={dataApiKey}
     bundlerApiKey={bundlerApiKey}
+    isModular={accountTemp === 'etherspotModular'}
   >
     <EtherspotTransactionKitContextProvider>
       <ProviderWalletContextProvider>
@@ -38,6 +60,7 @@ const EtherspotTransactionKit = ({
       </ProviderWalletContextProvider>
     </EtherspotTransactionKitContextProvider>
   </EtherspotContextProvider>
-);
+  )
+};
 
 export default EtherspotTransactionKit;